# contracts/openapi/designs.yaml
openapi: 3.1.0
info:
  title: Service designs — modèles et patrons
  version: 1.0.0
  description: >
    Modèles, versions de modèle et spécifications de patron. Le service est le seul à écrire
    dans sa base ; il publie design.versioned à chaque nouvelle version.
servers:
  - url: http://localhost:3101
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        '200':
          description: Le service répond.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Health' }
  /v1/designs:
    post:
      operationId: createDesign
      summary: Créer un modèle
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDesignRequest' }
      responses:
        '201':
          description: Modèle créé.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Design' }
        '400': { $ref: '#/components/responses/Problem' }
  /v1/designs/{designId}:
    parameters:
      - $ref: '#/components/parameters/DesignId'
    get:
      operationId: getDesign
      summary: Lire un modèle
      responses:
        '200':
          description: Le modèle.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Design' }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/designs/{designId}/versions:
    parameters:
      - $ref: '#/components/parameters/DesignId'
    post:
      operationId: createDesignVersion
      summary: Créer une version (calcule le patron)
      description: >
        Appelle le moteur de patronage avec les mesures et les paramètres, puis enregistre la
        version et sa spécification de patron. Mêmes entrées, même empreinte.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDesignVersionRequest' }
      responses:
        '201':
          description: Version créée.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DesignVersion' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '502': { $ref: '#/components/responses/Problem' }
  /v1/designs/{designId}/versions/{versionNumber}:
    parameters:
      - $ref: '#/components/parameters/DesignId'
      - name: versionNumber
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    get:
      operationId: getDesignVersion
      summary: Lire une version
      responses:
        '200':
          description: La version et son patron.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DesignVersion' }
        '404': { $ref: '#/components/responses/Problem' }
components:
  parameters:
    DesignId:
      name: designId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Problem:
      description: Erreur au format RFC 9457.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok] }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, description: 'Identifiant stable, ex. /problems/design-not-found' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
    CreateDesignRequest:
      $ref: '../schemas/designs/create-design-request.schema.json'
    Design:
      $ref: '../schemas/designs/design.schema.json'
    CreateDesignVersionRequest:
      $ref: '../schemas/designs/create-design-version-request.schema.json'
    DesignVersion:
      $ref: '../schemas/designs/design-version.schema.json'